requirements.txt
google-cloud-functions==1.16.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.1.2
requests==2.32.3
tweepy==4.14.0
